Fix from and to (Rev) handling if only one of this values is available.
The idea is to map null values to a special string that will be mapped
the the first/last revision when generating the list of revisions to
process. By design, the actual values of 'from' and 'to' are not
available after creating an AbstractIntervalVCSEngine. Thus, I removed
the public getter methods to avoid issues.